Fix stock showing as concatenated string instead of a sum
sum() over an int column returns bigint in Postgres, which node-postgres
returns as a string — that made the admin panel's total stock render as
"0393996" instead of adding 393 + 996. Cast back to int.

+++ src/app/drivers/repositories/products/SQLProductsRepository.ts
@@ -35,7 +35,9 @@ export class SQLProductsRepository implements IProductsRepository {
   private static readonly BASE_SELECT = `
     select
       p.id, p.slug, p.sku, p.name, p.price, p.currency,
-      coalesce(s.total_stock, 0) as stock,
+      -- sum() over int returns bigint, which node-postgres returns as a
+      -- string — cast back to int so callers get a real number.
+      coalesce(s.total_stock, 0)::int as stock,
       p.is_active as "isActive",
       p.is_internal as "isInternal",
       p.bundle_units as "bundleUnits",
